refactor(admin): give the rail sections one canonical vocabulary
The sidebar spelled its category union inline, so a route registry naming the
same values would have restated them. Both now derive from one declaration,
and the declarable set is separated from the resolvable one: a standalone
plugin's id exists only while that plugin is mounted, so it cannot be written
as a static literal without admitting an id that matches nothing.
